Refactor: Move cliutils to commands._cliutils
Moved all implementations from `src/dictknife/cliutils` to `src/dictknife/commands/_cliutils.py`.
Updated all import paths accordingly.
